requirements.txt
azure-core==1.30.1
azure-devops==7.1.0b4
alembic==1.13.1
SQLAlchemy==1.3.13
psycopg2==2.9.9
fastapi[all]
loguru==0.7.2
aiohttp==3.9.5
psycopg2==2.9.9
psycopg2-binary==2.9.9